Txpool synchronization and distributing package to files.
