+++ .vscode/config.json
@@ -1,0 +1,4 @@
+{
+    "node1_ipc": "/data/data_Tomer/node1/geth.ipc",
+    "node2_ipc": "/data/data_Tomer/node2/geth.ipc"
+}
