+++ riblt_with_certainty/ibf.go
@@ -1,74 +1,8 @@
 package riblt_with_certainty
 
 import (
-	"crypto/sha256"
-
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/holiman/uint256"
-	"github.com/kavehmz/prime"
 )
-
-var primes []uint64
-
-func init() {
-	primes = prime.Primes(1000000)
-}
-
-// Symbol is an interface that can be either a
-// common.Hash or a uint64
-type Symbol interface {
-	Xor(Symbol) Symbol
-	Hash() [32]byte
-	IsZero() bool
-	Equal(Symbol) bool
-}
-
-// HashSymbol wraps common.Hash to implement the Symbol interface
-type HashSymbol common.Hash
-
-func (h HashSymbol) Xor(other Symbol) Symbol {
-	o := other.(HashSymbol)
-	var result HashSymbol
-	for i := 0; i < common.HashLength; i++ {
-		result[i] = h[i] ^ o[i]
-	}
-	return result
-}
-
-func (h HashSymbol) Hash() [32]byte {
-	return sha256.Sum256(h[:])
-}
-
-func (h HashSymbol) IsZero() bool {
-	return h == HashSymbol{}
-}
-
-func (h HashSymbol) Equal(other Symbol) bool {
-	o := other.(HashSymbol)
-	return h == o
-}
-
-// Uint64Symbol wraps uint64 to implement the Symbol interface
-type Uint64Symbol uint64
-
-func (u Uint64Symbol) Xor(other Symbol) Symbol {
-	o := other.(Uint64Symbol)
-	return Uint64Symbol(uint64(u) ^ uint64(o))
-}
-
-func (u Uint64Symbol) Hash() [32]byte {
-	buf := uint256.NewInt(uint64(u)).Bytes()
-	return sha256.Sum256(buf)
-}
-
-func (u Uint64Symbol) IsZero() bool {
-	return u == 0
-}
-
-func (u Uint64Symbol) Equal(other Symbol) bool {
-	o := other.(Uint64Symbol)
-	return u == o
-}
 
 type IBFCell struct {
 	Count   int64
@@ -136,23 +70,6 @@ func (c *IBFCell) IsZero() bool {
 	return c.Count == 0 && c.XorSum.IsZero() && isZeroBytes(c.HashSum)
 }
 
-func xorBytes(a, b [32]byte) [32]byte {
-	var result [32]byte
-	for i := 0; i < 32; i++ {
-		result[i] = a[i] ^ b[i]
-	}
-	return result
-}
-
-func isZeroBytes(a [32]byte) bool {
-	for _, b := range a {
-		if b != 0 {
-			return false
-		}
-	}
-	return true
-}
-
 func (ibf *InvertibleBloomFilter) AddSymbols(symbols []Symbol) {
 	curPrime := primes[ibf.Iteration]
 
@@ -190,21 +107,6 @@ func (ibf *InvertibleBloomFilter) AddSymbols(symbols []Symbol) {
 	ibf.Iteration++
 }
 
-func ModSymbolUint64(s Symbol, m uint64) uint64 {
-	switch sym := s.(type) {
-	case HashSymbol:
-		var bigIntHash *uint256.Int
-		bigIntHash.SetBytes(sym[:])
-		bigIntMod := uint256.NewInt(m)
-		bigIntHash.Mod(bigIntHash, bigIntMod)
-		return bigIntHash.Uint64()
-	case Uint64Symbol:
-		return uint64(sym) % m
-	default:
-		panic("Unsupported Symbol type")
-	}
-}
-
 func (ibf *InvertibleBloomFilter) Subtract(ibf2 *InvertibleBloomFilter) *InvertibleBloomFilter {
 	difference := NewIBF(ibf.Size, ibf.SymbolType)
 	difference.Copy(ibf)
@@ -216,7 +118,7 @@ func (ibf *InvertibleBloomFilter) Subtract(ibf2 *InvertibleBloomFilter) *Inverti
 	return difference
 }
 
-func (ibf *InvertibleBloomFilter) Decode() (aWithoutB []Symbol, bWithoutA []Symbol, ok bool) {
+func (ibf *InvertibleBloomFilter) Decode() (symmetricDiff []Symbol, ok bool) {
 	pureList := make([]uint64, 0)
 
 	for {
@@ -244,14 +146,9 @@ func (ibf *InvertibleBloomFilter) Decode() (aWithoutB []Symbol, bWithoutA []Symb
 		count := ibf.Cells[j].Count
 		xorSum := ibf.Cells[j].XorSum
 
-		if count > 0 {
-			aWithoutB = append(aWithoutB, xorSum)
-		} else {
-			bWithoutA = append(bWithoutA, xorSum)
-		}
+		symmetricDiff = append(symmetricDiff, xorSum)
 
 		primesSum := uint64(0)
-
 		for _, prime := range primes[:ibf.Iteration] {
 			cellIdx := primesSum + ModSymbolUint64(xorSum, prime)
 			ibf.Cells[cellIdx].Count -= count
