+++ riblt_with_certainty/success_rate_vs_total_bits_test.go
@@ -58,7 +58,7 @@ func runTrialSuccessRateVsTotalCells(trialNumber int, universeSize int, symmetri
 
 		// Subtract the two IBFs and Decode the result to find the differences
 		ibfDiff := ibfBob.Subtract(ibfAlice)
-		bobWithoutAlice, _, _ := ibfDiff.Decode()
+		bobWithoutAlice, _ := ibfDiff.Decode()
 
 		if len(bobWithoutAlice) > 0 {
 			curSymmetricDiffSize = len(bobWithoutAlice)
